legacy/aws/lambda_orchestrator/requirements.txt
# AWS Strands SDK for multi-agent orchestration
strands-agents==0.1.6
strands-agents-tools==0.1.4

# AWS SDK for Lambda runtime and services
boto3==1.38.32
botocore==1.38.32

# HTTP requests for GitHub API integration  
requests==2.32.3

# Note: CodeRipple package installed separately via Terraform
